+++ src/lib/cron/core/index.test.ts
@@ -1,6 +1,8 @@
-import { describe, expect, test } from "bun:test";
-import { CronLengthError, OutOfBoundError } from "../errors.js";
-import { Cron } from "./index.js";
+import { describe, expect, setSystemTime, spyOn, test } from "bun:test";
+import { mightThrow, mightThrowSync } from "../../errors/index.js";
+import { Cron, RetryCronJob } from "./index.js";
+
+class UserDefinedError extends Error {}
 
 describe("Cron", () => {
   describe("constructor", () => {
@@ -27,6 +29,7 @@ describe("Cron", () => {
       });
 
       expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
     });
 
     test("should create a cron job with @hourly alias", () => {
@@ -39,9 +42,10 @@ describe("Cron", () => {
       });
 
       expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
     });
 
-    test("should throw error for invalid cron expression", () => {
+    test("should throw an error for invalid cron expression", () => {
       const handler = () => Promise.resolve();
 
       expect(() => {
@@ -49,11 +53,11 @@ describe("Cron", () => {
           name: "invalid-job",
           schedule: "invalid",
           handler,
-        });
-      }).toThrow(CronLengthError);
+        }).run();
+      }).toThrow(TypeError);
     });
 
-    test("should throw error for cron expression with wrong field count", () => {
+    test("should throw an error for cron expression with wrong field count", () => {
       const handler = () => Promise.resolve();
 
       expect(() => {
@@ -61,178 +65,36 @@ describe("Cron", () => {
           name: "invalid-job",
           schedule: "0 0 * *",
           handler,
-        });
-      }).toThrow(CronLengthError);
-    });
-  });
-
-  describe("parsing", () => {
-    test("should throw error for cron expression with a very large number", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "10000000 0 * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
+        }).run();
+      }).toThrow(TypeError);
     });
 
-    test("should reject an out-of-bounds number in second (100)", () => {
+    test("should throw an error when passing an invalid expression", () => {
       const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "100 * * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should reject an out-of-bounds step range in minute field (50-70/5)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "* 50-70/5  * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should accept a step range in minute field without the starting value", () => {
-      const handler = () => Promise.resolve();
-
       const cron = new Cron({
         name: "invalid-job",
-        schedule: "* -10/5  * * * *",
+        schedule: "0 0 * *",
         handler,
       });
 
+      const [scheduleFormatError] = mightThrowSync(() => cron.run());
       expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
+      expect(scheduleFormatError).toBeDefined();
     });
 
-    test("should accept a step range in minute field without the final value", () => {
+    test("should throw an error when passing an invalid alias", () => {
       const handler = () => Promise.resolve();
-
       const cron = new Cron({
-        name: "invalid-job",
-        schedule: "* 10/5  * * * *",
+        name: "invalid-alias",
+        schedule: "@invalid-alias",
         handler,
       });
 
+      const [invalidAliasError] = mightThrowSync(() => cron.run());
       expect(cron).toBeDefined();
-    });
-
-    test("should reject an out-of-bounds step range in minute field (-70/5)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "* -70/5  * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should reject an out-of-bounds step range in minute field (70/5)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "* 70/5  * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should reject an out-of-bounds step range in day of the week field (7)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "0 0 * * 7",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should reject an out-of-bounds list value in hour field (1,15,40)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "0 1,15,40 * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should reject an out-of-bounds range value in day field (1-33)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "0 * 1-33 * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("inverted range with step inside a list passes scanner", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "30-10/5 * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("inverted range inside a list passes scanner", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "30-10 * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("wildcard-with-step = 0 should fail", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "*/0,30 * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
-    });
-
-    test("should reject an out-of-bounds step range in minute field (2-70/5)", () => {
-      const handler = () => Promise.resolve();
-
-      expect(() => {
-        new Cron({
-          name: "invalid-job",
-          schedule: "2-70/5 * * * *",
-          handler,
-        });
-      }).toThrow(OutOfBoundError);
+      expect(cron.getStatus()).toBe("idle");
+      expect(invalidAliasError).toBeDefined();
     });
   });
 
@@ -248,53 +110,14 @@ describe("Cron", () => {
 
       expect(cron.getStatus()).toBe("idle");
 
-      cron.start();
+      cron.run();
       expect(cron.getStatus()).toBe("running");
 
       cron.stop();
       expect(cron.getStatus()).toBe("idle");
     });
 
-    test("should pause and resume a cron job", () => {
-      const handler = () => Promise.resolve();
-
-      const cron = new Cron({
-        name: "pause-test",
-        schedule: "0 0 * * *",
-        handler,
-      });
-
-      cron.start();
-      expect(cron.getStatus()).toBe("running");
-
-      cron.pause();
-      expect(cron.getStatus()).toBe("paused");
-
-      cron.resume();
-      expect(cron.getStatus()).toBe("running");
-
-      cron.stop();
-    });
-
-    test("should not resume if not paused", () => {
-      const handler = () => Promise.resolve();
-
-      const cron = new Cron({
-        name: "resume-test",
-        schedule: "0 0 * * *",
-        handler,
-      });
-
-      cron.start();
-      expect(cron.getStatus()).toBe("running");
-
-      cron.resume();
-      expect(cron.getStatus()).toBe("running");
-
-      cron.stop();
-    });
-
-    test("should not pause if not running", () => {
+    test("should not stop if not running", () => {
       const handler = () => Promise.resolve();
 
       const cron = new Cron({
@@ -304,11 +127,11 @@ describe("Cron", () => {
       });
 
       expect(cron.getStatus()).toBe("idle");
-      cron.pause();
+      cron.stop();
       expect(cron.getStatus()).toBe("idle");
     });
 
-    test("should handle multiple start calls gracefully", () => {
+    test("should handle multiple start calls gracefully without creating multiple jobs", () => {
       const handler = () => Promise.resolve();
 
       const cron = new Cron({
@@ -317,90 +140,120 @@ describe("Cron", () => {
         handler,
       });
 
-      cron.start();
-      cron.start();
-      cron.start();
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
 
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
+
+      cron.run();
       expect(cron.getStatus()).toBe("running");
 
       cron.stop();
+      expect(cron.getStatus()).toBe("idle");
     });
   });
 
   describe("aliases", () => {
-    test("should support @yearly alias", () => {
-      expect(() => {
-        new Cron({
-          name: "yearly",
-          schedule: "@yearly",
-          handler: () => Promise.resolve(),
-        });
-      }).not.toThrow();
+    test("should support @yearly and @annually alias", () => {
+      const yearly = new Cron({
+        name: "yearly",
+        schedule: "@yearly",
+        handler: () => Promise.resolve(),
+      });
+
+      const annually = new Cron({
+        name: "yearly",
+        schedule: "@annually",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => yearly.run()).not.toThrow();
+      expect(() => annually.run()).not.toThrow();
+
+      yearly.stop();
+      annually.stop();
     });
 
     test("should support @monthly alias", () => {
-      expect(() => {
-        new Cron({
-          name: "monthly",
-          schedule: "@monthly",
-          handler: () => Promise.resolve(),
-        });
-      }).not.toThrow();
+      const monthly = new Cron({
+        name: "monthly",
+        schedule: "@monthly",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => monthly.run()).not.toThrow();
+      monthly.stop();
     });
 
     test("should support @weekly alias", () => {
-      expect(() => {
-        new Cron({
-          name: "weekly",
-          schedule: "@weekly",
-          handler: () => Promise.resolve(),
-        });
-      }).not.toThrow();
+      const weekly = new Cron({
+        name: "weekly",
+        schedule: "@weekly",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => weekly.run()).not.toThrow();
+      weekly.stop();
     });
 
-    test("should support @daily alias", () => {
-      expect(() => {
-        new Cron({
-          name: "daily",
-          schedule: "@daily",
-          handler: () => Promise.resolve(),
-        });
-      }).not.toThrow();
+    test("should support @daily and @midnight alias", () => {
+      const daily = new Cron({
+        name: "daily",
+        schedule: "@daily",
+        handler: () => Promise.resolve(),
+      });
+
+      const midnight = new Cron({
+        name: "midnight",
+        schedule: "@midnight",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => daily.run()).not.toThrow();
+      expect(() => midnight.run()).not.toThrow();
+
+      daily.stop();
+      midnight.stop();
     });
 
     test("should support @hourly alias", () => {
-      expect(() => {
-        new Cron({
-          name: "hourly",
-          schedule: "@hourly",
-          handler: () => Promise.resolve(),
-        });
-      }).not.toThrow();
+      const hourly = new Cron({
+        name: "hourly",
+        schedule: "@hourly",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => hourly.run()).not.toThrow();
+      hourly.stop();
     });
 
     test("should support @minutely alias", () => {
-      expect(() => {
-        new Cron({
-          name: "minutely",
-          schedule: "@minutely",
-          handler: () => Promise.resolve(),
-        });
-      }).not.toThrow();
+      const minutely = new Cron({
+        name: "minutely",
+        schedule: "@minutely",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => minutely.run()).not.toThrow();
+      minutely.stop();
     });
   });
 
-  describe("getNextRun search horizon", () => {
+  describe("next() search horizon", () => {
     test("should find next run for yearly schedule", () => {
+      setSystemTime();
+
       const cron = new Cron({
         name: "yearly-horizon",
         schedule: "@yearly",
         handler: () => Promise.resolve(),
       });
 
-      const next = cron.getNextRun();
-
+      const next = cron.next();
       if (!next) throw new Error("Expected next run to be found");
 
+      expect(next.getFullYear()).toBe(new Date().getFullYear() + 1);
       expect(next.getMonth()).toBe(0);
       expect(next.getDate()).toBe(1);
       expect(next.getHours()).toBe(0);
@@ -415,8 +268,7 @@ describe("Cron", () => {
         handler: () => Promise.resolve(),
       });
 
-      const next = cron.getNextRun();
-
+      const next = cron.next();
       if (!next) throw new Error("Expected next run to be found");
 
       expect(next.getMonth()).toBe(11);
@@ -430,66 +282,105 @@ describe("Cron", () => {
       const cron = new Cron({
         name: "leap-day",
         schedule: "0 0 29 2 *",
-        handler: () => {},
+        handler: () => Promise.resolve(),
       });
 
-      const next = cron.getNextRun();
-
+      const next = cron.next();
       expect(next).not.toBeNull();
     });
-  });
 
-  describe("next retry on null getNextRun", () => {
-    test("should schedule a retry timeout when getNextRun returns null", () => {
+    test("should find next run when using a past starting date", () => {
       const cron = new Cron({
-        name: "retry-test",
-        schedule: "0 0 * * *",
+        name: "daily-horizon-from-the-past",
+        schedule: "@daily",
         handler: () => Promise.resolve(),
       });
 
-      cron.getNextRun = () => null;
+      const next = cron.next(new Date(2026, 4, 8));
+      if (!next) throw new Error("Expected next run to be found");
 
-      cron.start();
-
-      // Status should remain running (not stuck idle)
-      expect(cron.getStatus()).toBe("running");
-
-      // A timeout should have been set (stop clears it, proving it exists)
-      cron.stop();
-      expect(cron.getStatus()).toBe("idle");
+      expect(next.getFullYear()).toBe(2026);
+      expect(next.getMonth()).toBe(4);
+      expect(next.getDate()).toBe(9);
     });
 
-    test("should allow pause to clear retry timeout", () => {
+    test("should find next run when using a future starting date", () => {
       const cron = new Cron({
-        name: "retry-pause",
-        schedule: "0 0 * * *",
+        name: "daily-horizon-from-the-future",
+        schedule: "@daily",
         handler: () => Promise.resolve(),
       });
 
-      cron.getNextRun = () => null;
+      const next = cron.next(new Date(2027, 4, 8));
+      if (!next) throw new Error("Expected next run to be found");
 
-      cron.start();
-      expect(cron.getStatus()).toBe("running");
-
-      cron.pause();
-      expect(cron.getStatus()).toBe("paused");
+      expect(next.getFullYear()).toBe(2027);
+      expect(next.getMonth()).toBe(4);
+      expect(next.getDate()).toBe(9);
     });
 
-    test("should keep status running during retry, not idle", () => {
+    test("should find next run when using the @minutely alias", () => {
       const cron = new Cron({
-        name: "retry-status",
-        schedule: "0 0 * * *",
+        name: "minutely-horizon",
+        schedule: "@minutely",
         handler: () => Promise.resolve(),
       });
 
-      cron.getNextRun = () => null;
+      const next = cron.next(new Date(2027, 4, 8, 0, 16, 0));
+      if (!next) throw new Error("Expected next run to be found");
 
-      cron.start();
+      expect(next.getFullYear()).toBe(2027);
+      expect(next.getMonth()).toBe(4);
+      expect(next.getDate()).toBe(8);
+      expect(next.getHours()).toBe(0);
+      expect(next.getMinutes()).toBe(17);
+    });
 
-      // Must stay running, not fall back to idle
-      expect(cron.getStatus()).toBe("running");
+    test("should return null when no match is found", () => {
+      const cron = new Cron({
+        name: "leap-day",
+        schedule: "0 0 31 2 *",
+        handler: () => Promise.resolve(),
+      });
 
-      cron.stop();
+      const next = cron.next();
+      expect(next).toBeNull();
+    });
+
+    test("should raise a TypeError when starting date an invalid number", () => {
+      const cron = new Cron({
+        name: "leap-day",
+        schedule: "0 0 2 2 *",
+        handler: () => Promise.resolve(),
+      });
+
+      expect(() => cron.next(NaN)).toThrow(TypeError);
+      expect(() => cron.next(Infinity)).toThrow(TypeError);
+      expect(() => cron.next(-Infinity)).toThrow(TypeError);
+    });
+
+    test("should raise an error when the received starting date is invalid", () => {
+      const cron = new Cron({
+        name: "on-error-call",
+        schedule: "0 0 18 2 *",
+        handler: () => Promise.resolve(),
+      });
+
+      const [nanError, nanDate] = mightThrowSync(() => cron.next(NaN));
+      const [infinityError, infinityDate] = mightThrowSync(() =>
+        cron.next(Infinity),
+      );
+      const [negativeInfinityError, negativeInfinityDate] = mightThrowSync(() =>
+        cron.next(-Infinity),
+      );
+
+      expect(nanDate).toBeNull();
+      expect(infinityDate).toBeNull();
+      expect(negativeInfinityDate).toBeNull();
+
+      expect(nanError).toBeDefined();
+      expect(infinityError).toBeDefined();
+      expect(negativeInfinityError).toBeDefined();
     });
   });
 
@@ -503,19 +394,10 @@ describe("Cron", () => {
 
       expect(cron.getStatus()).toBe("idle");
 
-      cron.start();
+      cron.run();
       expect(cron.getStatus()).toBe("running");
 
-      cron.pause();
-      expect(cron.getStatus()).toBe("paused");
-
-      cron.pause();
-      expect(cron.getStatus()).toBe("paused");
-
-      cron.resume();
-      expect(cron.getStatus()).toBe("running");
-
-      cron.resume();
+      cron.run();
       expect(cron.getStatus()).toBe("running");
 
       cron.stop();
@@ -526,22 +408,352 @@ describe("Cron", () => {
     });
   });
 
-  describe("matches", () => {
-    test("should use OR semantics when both day-of-month and day-of-week are restricted", () => {
-      // "0 0 15 * 1" = midnight on the 15th of any month OR any Monday
-      const cron = new Cron({
-        name: "test",
-        schedule: "0 0 15 * 1",
-        handler: () => {},
+  describe("Dispose API", () => {
+    test("should stop a job automatically", () => {
+      const job = new Cron({
+        name: "disposable",
+        schedule: "@minutely",
+        handler: async () => Promise.resolve(),
       });
 
-      // 2024-01-08 = Monday that is NOT the 15th - matches because it's Monday
-      const mondayNot15th = new Date("2024-01-08T00:00:00");
-      expect(cron.matches(mondayNot15th)).toBe(true);
+      const stopSpy = spyOn(job, "stop");
 
-      // 2024-02-15 = Thursday that IS the 15th - matches because it's the 15th
-      const fifteenthNotMonday = new Date("2024-02-15T00:00:00");
-      expect(cron.matches(fifteenthNotMonday)).toBe(true);
+      {
+        using scoped = job;
+        scoped.run();
+      }
+
+      expect(stopSpy).toHaveBeenCalledTimes(1);
+      expect(job.getStatus()).toBe("idle");
+    });
+  });
+
+  describe("Retry", () => {
+    test("should successfully call onError() callback", async () => {
+      setSystemTime(new Date("2020-01-10T00:00:00.000Z"));
+      const handler = () => Promise.resolve();
+
+      const retry = new RetryCronJob({
+        maxAttempts: 0,
+        onError: (err) => {
+          expect(err.error).toBeInstanceOf(Error);
+          expect(err.error.message).toBe("A generic error");
+          expect(err.failedAt).toBe(Date.now());
+          expect(err.name).toBe("retry-job");
+        },
+      });
+
+      const cron = new Cron({
+        name: "retry-job",
+        schedule: "0 0 * * *",
+        handler,
+        retry,
+      });
+
+      expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
+
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
+
+      await retry.update(cron, new Error("A generic error"));
+      expect(cron.getStatus()).toBe("idle");
+
+      setSystemTime();
+    });
+
+    test("should successfully throw an error when onError() callback is not defined", async () => {
+      const handler = () => Promise.resolve();
+      const retry = new RetryCronJob({
+        maxAttempts: 0,
+      });
+
+      const cron = new Cron({
+        name: "retry-job",
+        schedule: "0 0 * * *",
+        handler,
+        retry,
+      });
+
+      expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
+
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
+
+      const [theError] = await mightThrow(
+        retry.update(cron, new Error("A generic error")),
+      );
+
+      expect(cron.getStatus()).toBe("idle");
+      expect(theError).toBeDefined();
+      expect(theError?.message).toBe("A generic error");
+    });
+
+    test("should successfully call onFailedAttempt() callback", async () => {
+      const handler = () => Promise.resolve();
+
+      const retry = new RetryCronJob({
+        maxAttempts: 3,
+        onFailedAttempt: ({ attemptNumber, delay, error, retriesLeft }) => {
+          if (attemptNumber === 1) {
+            expect(attemptNumber).toBe(1);
+            expect(delay).toBeGreaterThan(0);
+            expect(error).toBeInstanceOf(Error);
+            expect(error.message).toBe("A generic error");
+            expect(retriesLeft).toBe(3);
+
+            return;
+          }
+
+          expect(attemptNumber).toBe(2);
+          expect(delay).toBeGreaterThan(0);
+          expect(error).toBeInstanceOf(Error);
+          expect(error.message).toBe("A generic error");
+          expect(retriesLeft).toBe(2);
+        },
+      });
+
+      const cron = new Cron({
+        name: "retry-job",
+        schedule: "0 0 * * *",
+        handler,
+        retry,
+      });
+
+      expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
+
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
+
+      await retry.update(cron, new Error("A generic error"));
+      expect(cron.getStatus()).toBe("running");
+
+      await retry.update(cron, new Error("A generic error"));
+      expect(cron.getStatus()).toBe("running");
+
+      cron.stop();
+      expect(cron.getStatus()).toBe("idle");
+    });
+
+    test("should not retry when retryOnError() callback return false", async () => {
+      const handler = () => Promise.resolve();
+
+      const retry = new RetryCronJob({
+        maxAttempts: 5,
+        retryOnError: (err) => !(err instanceof UserDefinedError),
+      });
+
+      const cron = new Cron({
+        name: "retry-job",
+        schedule: "0 0 * * *",
+        handler,
+        retry,
+      });
+
+      expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
+
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
+
+      await retry.update(cron, new Error("Generic error"));
+      expect(cron.getStatus()).toBe("running");
+
+      const [userDefinedError] = await mightThrow(
+        retry.update(cron, new UserDefinedError("User defined error")),
+      );
+
+      expect(cron.getStatus()).toBe("idle");
+      expect(userDefinedError).toBeDefined();
+      expect(userDefinedError?.message).toBe("User defined error");
+    });
+
+    test("should not retry when retryOnError() callback return false and also call onError() callback", async () => {
+      setSystemTime(new Date("2020-01-10T00:00:00.000Z"));
+      const handler = () => Promise.resolve();
+
+      const retry = new RetryCronJob({
+        maxAttempts: 5,
+        retryOnError: (err) => !(err instanceof UserDefinedError),
+        onError: (err) => {
+          expect(err.error).toBeInstanceOf(Error);
+          expect(err.error.message).toBe("User defined error");
+          expect(err.failedAt).toBe(Date.now());
+          expect(err.name).toBe("retry-job");
+        },
+      });
+
+      const cron = new Cron({
+        name: "retry-job",
+        schedule: "0 0 * * *",
+        handler,
+        retry,
+      });
+
+      expect(cron).toBeDefined();
+      expect(cron.getStatus()).toBe("idle");
+
+      cron.run();
+      expect(cron.getStatus()).toBe("running");
+
+      await retry.update(cron, new Error("Generic error"));
+      expect(cron.getStatus()).toBe("running");
+
+      await retry.update(cron, new UserDefinedError("User defined error"));
+      expect(cron.getStatus()).toBe("idle");
+
+      setSystemTime();
+    });
+
+    describe("Validate attempts", () => {
+      test("should raise an error when passing a negative number", async () => {
+        const handler = () => Promise.resolve();
+        const retryNegativeNumber = new RetryCronJob({
+          maxAttempts: -10,
+        });
+
+        const cron = new Cron({
+          name: "retry-job",
+          schedule: "0 0 * * *",
+          handler,
+        });
+
+        expect(cron).toBeDefined();
+        expect(cron.getStatus()).toBe("idle");
+
+        const [negativeNumberError] = await mightThrow(
+          retryNegativeNumber.update(cron, new Error("A generic error")),
+        );
+
+        expect(negativeNumberError).toBeDefined();
+        expect(negativeNumberError).toBeInstanceOf(TypeError);
+
+        const retryNegativeZero = new RetryCronJob({
+          maxAttempts: -0,
+        });
+
+        const [negativeZeroError] = await mightThrow(
+          retryNegativeZero.update(cron, new Error("A generic error")),
+        );
+
+        expect(negativeZeroError).toBeDefined();
+        expect(negativeZeroError).toBeInstanceOf(TypeError);
+
+        const retryNegativeInfinity = new RetryCronJob({
+          maxAttempts: Number.NEGATIVE_INFINITY,
+        });
+
+        const [negativeInfinityError] = await mightThrow(
+          retryNegativeInfinity.update(cron, new Error("A generic error")),
+        );
+
+        expect(negativeInfinityError).toBeDefined();
+        expect(negativeInfinityError).toBeInstanceOf(TypeError);
+      });
+
+      test("should raise an error when passing NaN", async () => {
+        const handler = () => Promise.resolve();
+        const retryNan = new RetryCronJob({
+          maxAttempts: NaN,
+        });
+
+        const cron = new Cron({
+          name: "retry-job",
+          schedule: "0 0 * * *",
+          handler,
+        });
+
+        expect(cron).toBeDefined();
+        expect(cron.getStatus()).toBe("idle");
+
+        const [nanError] = await mightThrow(
+          retryNan.update(cron, new Error("A generic error")),
+        );
+
+        expect(nanError).toBeDefined();
+        expect(nanError).toBeInstanceOf(TypeError);
+      });
+
+      test("should raise an error when passing a non-integer number", async () => {
+        const handler = () => Promise.resolve();
+        const decimalNumberRetry = new RetryCronJob({
+          maxAttempts: 1.5,
+        });
+
+        const cron = new Cron({
+          name: "retry-job",
+          schedule: "0 0 * * *",
+          handler,
+        });
+
+        expect(cron).toBeDefined();
+        expect(cron.getStatus()).toBe("idle");
+
+        const [decimalNumberError] = await mightThrow(
+          decimalNumberRetry.update(cron, new Error("A generic error")),
+        );
+
+        expect(decimalNumberError).toBeDefined();
+        expect(decimalNumberError).toBeInstanceOf(TypeError);
+      });
+
+      test("should not raise an error when passing a floating point number that can be represented as integer", async () => {
+        const handler = () => Promise.resolve();
+        const decimalNumberRetry = new RetryCronJob({
+          maxAttempts: 5.0,
+        });
+
+        const cron = new Cron({
+          name: "retry-job",
+          schedule: "0 0 * * *",
+          handler,
+        });
+
+        expect(cron).toBeDefined();
+        expect(cron.getStatus()).toBe("idle");
+
+        const [decimalNumberError] = await mightThrow(
+          decimalNumberRetry.update(cron, new Error("A generic error")),
+        );
+
+        expect(decimalNumberError).toBeNull();
+      });
+
+      test("should raise an error when passing 'Infinity'", async () => {
+        const handler = () => Promise.resolve();
+        const infinityRetry = new RetryCronJob({
+          maxAttempts: Number.POSITIVE_INFINITY,
+        });
+
+        const cron = new Cron({
+          name: "retry-job",
+          schedule: "0 0 * * *",
+          handler,
+        });
+
+        expect(cron).toBeDefined();
+        expect(cron.getStatus()).toBe("idle");
+
+        const [infinityError] = await mightThrow(
+          infinityRetry.update(cron, new Error("A generic error")),
+        );
+
+        expect(infinityError).toBeDefined();
+        expect(infinityError).toBeInstanceOf(TypeError);
+
+        const secondInfinityRetry = new RetryCronJob({
+          maxAttempts: Infinity,
+        });
+
+        const [secondInfinityError] = await mightThrow(
+          secondInfinityRetry.update(cron, new Error("A generic error")),
+        );
+
+        expect(secondInfinityError).toBeDefined();
+        expect(secondInfinityError).toBeInstanceOf(TypeError);
+      });
     });
   });
 });
